osh40-main/osh40-main/requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
charset-normalizer==3.0.1
comm==0.1.3
connectorx==0.3.1
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
decouple==0.0.7
defusedxml==0.7.1
Deprecated==1.2.13
executing==1.2.0
fastjsonschema==2.16.3
fonttools==4.39.3
fqdn==1.5.1
idna==3.4
importlib-metadata==6.5.0
importlib-resources==5.12.0
influxdb==5.3.1
influxdb-client==1.35.0
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
llvmlite==0.39.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
mqtt==0.0.1
msgpack==1.0.4
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.5
packaging==23.1
paho-mqtt==1.6.1
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.2.0
polars==0.15.18
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.7
pyarrow==11.0.0
pycparser==2.21
Pygments==2.15.1
pyparsing==3.0.9
pyrsistent==0.19.3
pyspark==2.4.8
python-dateutil==2.8.2
python-decouple==3.7
python-json-logger==2.0.7
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.2
reactivex==4.0.4
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0.post4
sktime==0.17.1
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
statsmodels==0.13.5
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tornado==6.3
traitlets==5.9.0
typing_extensions==4.4.0
uri-template==1.2.0
urllib3==1.26.14
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
wrapt==1.15.0
xgboost==1.7.6
zipp==3.15.0
